refactor: Minor cleanups
 - Add methods in TripSchedule to support itinerary mapping
 - Encapsulate `tripPatternForDates` and `offsets` in TripPatternForDates
 - Move all frequency based search and events in raptor to its own package
 - Make the events final
 - Reduce visibility where appropriate(from public to package local)
