+++ src/main/java/org/opentripplanner/transit/raptor/rangeraptor/transit/frequency/FrequencyBoardingEvent.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.transit.raptor.rangeraptor.transit;
+package org.opentripplanner.transit.raptor.rangeraptor.transit.frequency;
 
 import java.time.LocalDate;
 import org.opentripplanner.model.TripPattern;
@@ -6,9 +6,9 @@ import org.opentripplanner.routing.trippattern.TripTimes;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripPattern;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
 
-class FrequencyAlightEvent<T extends RaptorTripSchedule> extends FrequencyBoardOrAlightEvent<T> {
+final class FrequencyBoardingEvent<T extends RaptorTripSchedule> extends FrequencyBoardOrAlightEvent<T> {
 
-    public FrequencyAlightEvent(
+    public FrequencyBoardingEvent(
             RaptorTripPattern raptorTripPattern,
             TripTimes tripTimes,
             TripPattern pattern,
@@ -17,6 +17,7 @@ class FrequencyAlightEvent<T extends RaptorTripSchedule> extends FrequencyBoardO
             int headway,
             int offset,
             LocalDate serviceDate
+
     ) {
         super(
                 raptorTripPattern,
@@ -30,15 +31,14 @@ class FrequencyAlightEvent<T extends RaptorTripSchedule> extends FrequencyBoardO
         );
     }
 
+    // Add headway here to report a late enough time to account for uncertainty
     @Override
     public int arrival(int stopPosInPattern) {
-        return tripTimes.getArrivalTime(stopPosInPattern) + offset;
+        return tripTimes.getArrivalTime(stopPosInPattern) + headway + offset;
     }
 
-    // Remove headway here to report an early enough departure time for the raptor search
     @Override
     public int departure(int stopPosInPattern) {
-        return tripTimes.getDepartureTime(stopPosInPattern) - headway + offset;
+        return tripTimes.getDepartureTime(stopPosInPattern) + offset;
     }
-
 }
