+++ src/main/java/org/opentripplanner/transit/raptor/rangeraptor/transit/frequency/TripFrequencyBoardSearch.java
@@ -1,18 +1,18 @@
-package org.opentripplanner.transit.raptor.rangeraptor.transit;
+package org.opentripplanner.transit.raptor.rangeraptor.transit.frequency;
 
-import org.opentripplanner.routing.algorithm.raptor.transit.TripPatternForDate;
 import org.opentripplanner.routing.algorithm.raptor.transit.request.TripPatternForDates;
-import org.opentripplanner.routing.trippattern.FrequencyEntry;
 import org.opentripplanner.routing.trippattern.TripTimes;
+import org.opentripplanner.transit.raptor.api.transit.IntIterator;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTimeTable;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripScheduleBoardOrAlightEvent;
+import org.opentripplanner.transit.raptor.rangeraptor.transit.TripScheduleSearch;
 
-public final class TripFrequencyAlightSearch<T extends RaptorTripSchedule> implements TripScheduleSearch<T> {
+public final class TripFrequencyBoardSearch<T extends RaptorTripSchedule> implements TripScheduleSearch<T> {
 
     private final TripPatternForDates timeTable;
 
-    public TripFrequencyAlightSearch(RaptorTimeTable<T> timeTable) {
+    public TripFrequencyBoardSearch(RaptorTimeTable<T> timeTable) {
         this.timeTable = (TripPatternForDates) timeTable;
     }
 
@@ -22,30 +22,28 @@ public final class TripFrequencyAlightSearch<T extends RaptorTripSchedule> imple
             int stopPositionInPattern,
             int tripIndexLimit
     ) {
-        for (int i = timeTable.tripPatternForDates.length - 1; i >= 0; i--) {
-            TripPatternForDate pattern = timeTable.tripPatternForDates[i];
-            int offset = timeTable.offsets[i];
+        IntIterator indexIterator = timeTable.tripPatternForDatesIndexIterator(true);
+        while (indexIterator.hasNext()) {
+            int i = indexIterator.next();
+            var pattern = timeTable.tripPatternForDate(i);
+            int offset = timeTable.tripPatternForDateOffsets(i);
 
-            for (int j = pattern.getFrequencies().size() - 1; j >= 0; j--) {
-                final FrequencyEntry frequency = pattern.getFrequencies().get(j);
-                var arrivalTime = frequency.prevArrivalTime(
-                        stopPositionInPattern,
-                        earliestBoardTime - offset
-                );
-                if (arrivalTime != -1) {
+            for (var frequency : pattern.getFrequencies()) {
+                var departureTime = frequency.nextDepartureTime(stopPositionInPattern, earliestBoardTime - offset);
+                if (departureTime != -1) {
                     int headway = frequency.exactTimes ? 0 : frequency.headway;
                     TripTimes tripTimes = frequency.materialize(
                             stopPositionInPattern,
-                            arrivalTime + headway,
-                            false
+                            departureTime - headway,
+                            true
                     );
 
-                    return new FrequencyAlightEvent<>(
+                    return new FrequencyBoardingEvent<>(
                             timeTable,
                             tripTimes,
                             pattern.getTripPattern().getPattern(),
                             stopPositionInPattern,
-                            arrivalTime + headway,
+                            departureTime - headway,
                             headway,
                             offset,
                             pattern.getLocalDate()
